requirements.txt
appdirs==1.4.3
Django==1.11.1
django-markdown-deux==1.0.5
django-pagedown==0.1.3
markdown2==2.3.4
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0